CI: only run tests that might use threads for TSAN run
