+++ .github/workflows/compiler_sanitizers.yml
@@ -116,9 +116,8 @@ jobs:
         python -m spin build -j2 -- -Db_sanitize=thread
     - name: Test
       run: |
-        # pass -s to pytest to see TSAN errors and warnings, otherwise pytest captures them
+        # These tests are slow, so only run tests in files that do "import threading" to make them count
         TSAN_OPTIONS=allocator_may_return_null=1:halt_on_error=1 \
-        python -m spin test -- -v -s --timeout=600 --durations=10
-    - name: Setup tmate session
-      if: ${{ failure() }}
-      uses: mxschmitt/action-tmate@v3
+        python -m spin test \
+        `find numpy -name "test*.py" | xargs grep -l "import threading" | tr '\n' ' '` \
+        -- -v -s --timeout=600 --durations=10
